Add Accounting Features

+++ src/components/Home.js
@@ -4,8 +4,11 @@ import styled, { keyframes } from 'styled-components';
 import GlobalStyle from './GlobalStyle';
 import { IoHeart } from "react-icons/io5";
 import { GoPlus } from "react-icons/go";
-import { db } from '../firebase';
-import { collection, getDocs, doc, deleteDoc } from 'firebase/firestore';
+import { GoArrowUpRight } from "react-icons/go";
+import { IoLogOutOutline } from "react-icons/io5";
+import { db, auth } from '../firebase';
+import { collection, getDocs, doc, deleteDoc, query, where } from 'firebase/firestore';
+import { onAuthStateChanged, signOut } from 'firebase/auth';
 import Modal from './Modal';
 
 // 요소가 천천히 나타나는 애니메이션 정의
@@ -30,6 +33,17 @@ const IconPlus = styled(GoPlus)`
   font-size: 27px;
   vertical-align: middle;
 `;
+const IconLogin = styled(GoArrowUpRight)`
+  font-size: 15px;
+  vertical-align: middle;
+  margin-bottom: 3px;
+`;
+const IconLogout = styled(IoLogOutOutline)`
+  font-size: 15px;
+  vertical-align: middle;
+  margin-bottom: 3px;
+`;
+
 
 const Main = styled.div`
   margin: 0 auto;
@@ -44,7 +58,7 @@ const Main = styled.div`
 const Container = styled.div`
   position: relative;
   max-width: 450px;
-  width: 100%;
+  width: 100%;  
   min-height: 100vh;
   height: 100%;
   background-color: rgba(255, 255, 255, 0.7);
@@ -55,6 +69,7 @@ const Container = styled.div`
   box-sizing: border-box;
 `;
 const Header = styled.div`
+  margin-top: ${(props) => (props.shrunk ? "0px" : "25px")};
   max-width: 450px;
   width: 100%;
   height: ${(props) => (props.shrunk ? "30px" : "52px")};
@@ -79,9 +94,31 @@ const Title = styled.p`
   color: #323031;
   transition: height 0.5s ease;
 `;
+const Auth = styled.div`
+  max-width: 450px;
+  width: 100%;
+  height: 25px;
+  border-bottom: 1px solid #ededed;
+  background-color: #ffffff;
+  display: ${(props) => (props.shrunk ? "none" : "flex")};
+  align-items: center;
+  justify-content: center;
+  position: fixed;
+  z-index: 1000;
+`
+const SignInBtn = styled.button`
+  font-size: 15px;
+  text-align: center;
+  margin-left: 10px;
+  text-decoration: underline;
+`
+const LogoutBtn = styled.button`
+  position: absolute;
+  right: 7px;
+`
 const CardWrap = styled.div`
   animation: ${fadeIn} 1s ease-out;
-  padding: 70px 20px 30px 20px;
+  padding: 90px 20px 30px 20px;
 `;
 const Card = styled.div`
   padding: 12px;
@@ -190,6 +227,7 @@ const Footer = styled.div`
 export default function Home() {
   const [cafes, setCafes] = useState([]);
   const [selectedCafe, setSelectedCafe] = useState(null);
+  const [user, setUser] = useState(null);
 
   const handleCardClick = (cafe) => {
     setSelectedCafe(cafe);
@@ -200,29 +238,67 @@ export default function Home() {
   };
 
   useEffect(() => {
+    // 로그인 상태 체크
+    const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
+      if (currentUser) {
+        setUser(currentUser); // 로그인된 사용자 정보 설정
+      } else {
+        setUser(null); // 로그아웃 시 null로 설정
+      }
+    });
+
+    // 컴포넌트 언마운트 시 구독 해제
+    return () => unsubscribe();
+  },[]);
+
+  useEffect(() => {
     const fetchCafes = async () => {
-      const cafesCollection = collection(db, 'cafes');
-      const cafesSnapshot = await getDocs(cafesCollection);
-      let cafesList = cafesSnapshot.docs.map(doc => ({ id: doc.id, ...doc.data() }));
-      cafesList = cafesList.sort((a, b) => new Date(b.date) - new Date(a.date)); // 날짜 기준 내림차순
-      setCafes(cafesList);
+      if (user) {
+        // 로그인한 사용자의 데이터 가져오기
+        const cafesCollection = query(collection(db, 'cafes'), where('userId', '==', user.uid));
+        const cafesSnapshot = await getDocs(cafesCollection);
+        let cafesList = cafesSnapshot.docs.map(doc => ({ id: doc.id, ...doc.data() }));
+        cafesList = cafesList.sort((a, b) => new Date(b.date) - new Date(a.date)); // 날짜 기준 내림차순
+        setCafes(cafesList);
+      } else {
+        // 로그인하지 않은 사용자의 데이터 가져오기
+        const localList = JSON.parse(localStorage.getItem('cafes')) || [];
+        setCafes(localList);
+      }
     };
 
     fetchCafes();
-  }, []);
+  }, [user]);
 
   const handleDelete = async (id, e) => {
     e.stopPropagation(); // 이벤트 전파 방지(모달 열림 막기)
     const confirmed = window.confirm("정말 삭제하시나요?");
     if (confirmed) {
       try {
-        await deleteDoc(doc(db, 'cafes', id));
-        setCafes(prevCafes => prevCafes.filter(cafe => cafe.id !== id));
-        alert('삭제되었어요✔');
+        if (auth.currentUser) {
+          await deleteDoc(doc(db, 'cafes', id));
+          setCafes(prevCafes => prevCafes.filter(cafe => cafe.id !== id));
+          alert('삭제되었어요✔');
+        } else {
+          const localLists = JSON.parse(localStorage.getItem('cafes')) || [];
+          const updatedLists = localLists.filter(cafe => cafe.id !== parseInt(id));
+          localStorage.setItem('cafes', JSON.stringify(updatedLists));
+          setCafes(updatedLists);
+          alert('삭제되었어요 (로컬)✔');
+        }
         setSelectedCafe(null);
       } catch (e) {
         console.error('삭제 중 오류 발생: ', e);
       }
+    }
+  };
+
+  const handleLogout = async () => {
+    try {
+      await signOut(auth);
+      /* navigate('/'); */
+    } catch (e) {
+      console.error('로그아웃 중 오류 발생: ', e);
     }
   };
 
@@ -246,30 +322,67 @@ export default function Home() {
     <Main>
       <GlobalStyle />
       <Container>
+        {user ? (
+          <Auth>
+            <p style={{fontFamily: 'Pretendard-Regular', fontSize: '12px'}}>{user.email}</p>
+            <LogoutBtn onClick={handleLogout}><IconLogout/></LogoutBtn>
+          </Auth>
+        ) : (
+          <Auth shrunk={isShrunk}>
+            <p>로그인하고 안전하게 이용하세요.</p>
+            <Link to="/login" style={{all: 'unset'}}><SignInBtn>로그인<IconLogin /></SignInBtn></Link>
+          </Auth>
+        )}
         <Header shrunk={isShrunk}>
           <Title shrunk={isShrunk}>컵일기.</Title>
         </Header>
         {cafes.length > 0 ? (
           <CardWrap>
-          {cafes.map((cafe, index) => (
-            <Card onClick={() => handleCardClick(cafe)} key={index}>
-              <CardMain>
-                {cafe.imageUrl && <Img src={cafe.imageUrl} alt="Cafe" />}
-                <ImgText>
-                  <Name>{cafe.cafeName}</Name>
-                  <Rating>
-                    {[...Array(5)].map((_, i) => (
-                      <IconHeart key={i} filled={i < cafe.rating ? 1 : 0} />
-                    ))}
-                  </Rating>
-                </ImgText>
-              </CardMain>
-              <CardText>{cafe.text}</CardText>
-              <CardFooter>
-                <DateArea>{new Date(cafe.date).toLocaleDateString('ko-KR', { year: 'numeric', month: 'long', day: 'numeric' })}</DateArea>
-              </CardFooter>
-            </Card>
-          ))}
+            {user ? (
+              <div>
+                {cafes.map((cafe) => (
+                  <Card onClick={() => handleCardClick(cafe)} key={cafe.id}>
+                    <CardMain>
+                      {cafe.imageUrl && <Img src={cafe.imageUrl} alt="Cafe" />}
+                      <ImgText>
+                        <Name>{cafe.cafeName}</Name>
+                        <Rating>
+                          {[...Array(5)].map((_, i) => (
+                            <IconHeart key={i} filled={i < cafe.rating ? 1 : 0} />
+                          ))}
+                        </Rating>
+                      </ImgText>
+                    </CardMain>
+                    <CardText>{cafe.text}</CardText>
+                    <CardFooter>
+                      <DateArea>{new Date(cafe.date).toLocaleDateString('ko-KR', { year: 'numeric', month: 'long', day: 'numeric' })}</DateArea>
+                    </CardFooter>
+                  </Card>
+                ))}
+              </div>
+            ) : (
+              <div>
+                {cafes.map((cafe, index) => (
+                  <Card onClick={() => handleCardClick(cafe)} key={index}>
+                    <CardMain>
+                      {cafe.imageUrl && <Img src={cafe.imageUrl} alt="Cafe" />}
+                      <ImgText>
+                        <Name>{cafe.cafeName}</Name>
+                        <Rating>
+                          {[...Array(5)].map((_, i) => (
+                            <IconHeart key={i} filled={i < cafe.rating ? 1 : 0} />
+                          ))}
+                        </Rating>
+                      </ImgText>
+                    </CardMain>
+                    <CardText>{cafe.text}</CardText>
+                    <CardFooter>
+                      <DateArea>{new Date(cafe.date).toLocaleDateString('ko-KR', { year: 'numeric', month: 'long', day: 'numeric' })}</DateArea>
+                    </CardFooter>
+                  </Card>
+                ))}
+              </div>
+            )}
           </CardWrap>
         ) : (
           <Text>아직 기록이 없어요. <br/>다녀온 카페를 추가해 볼까요?</Text>
